Add a love2d visualization of the easing functions
This demo uses love2d, but of course flux still has no dependencies on
love2d.

To make it easier to pick which function to use, visualize them in four
ways: a plot, translation, alpha, and scale.

This is similar to easings.net, but uses the same names as flux, only
shows the functions provided by flux, and actually uses flux for
rendering so it's definitely accurate.